remasterizacion base de datos e software remasterizacion 60%

+++ SisNikosPizza.Domain/Models/DetallePedido.cs
@@ -10,7 +10,7 @@ namespace SisNikosPizza.Domain.Models
         public int ProductoId { get; set; }
         public Producto? Producto { get; set; }
         public int Cantidad { get; set; }
-        public int PrecioUnitario { get; set; }
-        public float PrecioTotal { get; set; }
+        public decimal PrecioUnitario { get; set; }
+        public decimal PrecioTotal { get; set; }
     }
 }
